calculator: changed base package from steammachinist to ru.neoflex.calculator
