+++ calculator/src/main/java/ru/neoflex/calculator/CalculatorApplication.java
@@ -1,0 +1,12 @@
+package ru.neoflex.calculator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CalculatorApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CalculatorApplication.class, args);
+    }
+}
