+++ calculator/src/main/java/ru/neoflex/calculator/dto/exception/CreditDeniedException.java
@@ -1,0 +1,8 @@
+package ru.neoflex.calculator.dto.exception;
+
+public class CreditDeniedException extends RuntimeException {
+
+    public CreditDeniedException(String message) {
+        super(message);
+    }
+}
