+++ calculator/src/main/java/ru/neoflex/calculator/dto/scoring/request/EmploymentStatus.java
@@ -1,0 +1,9 @@
+package ru.neoflex.calculator.dto.scoring.request;
+
+public enum EmploymentStatus {
+
+    UNEMPLOYED,
+    SELF_EMPLOYED,
+    EMPLOYED,
+    BUSINESS_OWNER
+}
