+++ calculator/src/main/java/ru/neoflex/calculator/dto/scoring/request/Gender.java
@@ -1,0 +1,8 @@
+package ru.neoflex.calculator.dto.scoring.request;
+
+public enum Gender {
+
+    MALE,
+    FEMALE,
+    NON_BINARY
+}
