+++ calculator/src/main/java/ru/neoflex/calculator/dto/scoring/request/MaritalStatus.java
@@ -1,0 +1,9 @@
+package ru.neoflex.calculator.dto.scoring.request;
+
+public enum MaritalStatus {
+
+    MARRIED,
+    DIVORCED,
+    SINGLE,
+    WIDOW_WIDOWER
+}
