+++ calculator/src/main/java/ru/neoflex/calculator/dto/scoring/request/Position.java
@@ -1,0 +1,9 @@
+package ru.neoflex.calculator.dto.scoring.request;
+
+public enum Position {
+
+    WORKER,
+    MID_MANAGER,
+    TOP_MANAGER,
+    OWNER
+}
