+++ calculator/src/main/java/ru/neoflex/calculator/util/CalculatorUtil.java
@@ -1,0 +1,18 @@
+package ru.neoflex.calculator.util;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CalculatorUtil {
+    public static List<boolean[]> generateBooleanCombinations() {
+        List<boolean[]> combinations = new ArrayList<>();
+
+        boolean[] values = {false, true};
+        for (boolean first : values) {
+            for (boolean second : values) {
+                combinations.add(new boolean[]{first, second});
+            }
+        }
+        return combinations;
+    }
+}
