+++ calculator/src/main/java/ru/neoflex/calculator/validation/Adult.java
@@ -1,0 +1,18 @@
+package ru.neoflex.calculator.validation;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER, ElementType.ANNOTATION_TYPE })
+@Retention(RetentionPolicy.RUNTIME)
+@Constraint(validatedBy = { AdultValidator.class })
+public @interface Adult {
+
+    String message() default "The user must be 18 years or older";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
