+++ calculator/src/main/java/ru/neoflex/calculator/validation/AdultValidator.java
@@ -1,0 +1,21 @@
+package ru.neoflex.calculator.validation;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+import java.time.LocalDate;
+import java.time.Period;
+
+public class AdultValidator implements ConstraintValidator<Adult, LocalDate> {
+
+    @Override
+    public void initialize(Adult constraintAnnotation) {
+    }
+
+    @Override
+    public boolean isValid(LocalDate birthDate, ConstraintValidatorContext context) {
+        if (birthDate == null) {
+            return false;
+        }
+        return Period.between(birthDate, LocalDate.now()).getYears() >= 18;
+    }
+}
