--- calculator/src/main/java/steammachinist/CalculatorApplication.java
@@ -1,12 +1,0 @@
-package steammachinist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CalculatorApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CalculatorApplication.class, args);
-    }
-}
