--- calculator/src/main/java/steammachinist/dto/GeneralCreditInfo.java
@@ -1,30 +1,0 @@
-package steammachinist.dto;
-
-import steammachinist.dto.offer.response.LoanOfferDto;
-import steammachinist.dto.scoring.response.CreditDto;
-
-import java.math.BigDecimal;
-
-public interface GeneralCreditInfo {
-
-    default BigDecimal amount() {
-        if (this instanceof LoanOfferDto loanOffer) {
-            return loanOffer.requestedAmount();
-        } else if (this instanceof CreditDto credit) {
-            return credit.amount();
-        }
-        throw new UnsupportedOperationException(String.format("Unsupported type: %s", this.getClass().getName()));
-    }
-
-    Integer term();
-
-    BigDecimal monthlyPayment();
-    GeneralCreditInfo withMonthlyPayment(BigDecimal monthlyPayment);
-
-    BigDecimal rate();
-    GeneralCreditInfo withRate(BigDecimal rate);
-
-    Boolean isInsuranceEnabled();
-
-    Boolean isSalaryClient();
-}
