--- calculator/src/main/java/steammachinist/dto/exception/CreditDeniedException.java
@@ -1,8 +1,0 @@
-package steammachinist.dto.exception;
-
-public class CreditDeniedException extends RuntimeException {
-
-    public CreditDeniedException(String message) {
-        super(message);
-    }
-}
