--- calculator/src/main/java/steammachinist/dto/scoring/request/EmploymentStatus.java
@@ -1,9 +1,0 @@
-package steammachinist.dto.scoring.request;
-
-public enum EmploymentStatus {
-
-    UNEMPLOYED,
-    SELF_EMPLOYED,
-    EMPLOYED,
-    BUSINESS_OWNER
-}
