--- calculator/src/main/java/steammachinist/dto/scoring/request/Gender.java
@@ -1,8 +1,0 @@
-package steammachinist.dto.scoring.request;
-
-public enum Gender {
-
-    MALE,
-    FEMALE,
-    NON_BINARY
-}
