--- calculator/src/main/java/steammachinist/dto/scoring/request/MaritalStatus.java
@@ -1,9 +1,0 @@
-package steammachinist.dto.scoring.request;
-
-public enum MaritalStatus {
-
-    MARRIED,
-    DIVORCED,
-    SINGLE,
-    WIDOW_WIDOWER
-}
