--- calculator/src/main/java/steammachinist/dto/scoring/request/Position.java
@@ -1,9 +1,0 @@
-package steammachinist.dto.scoring.request;
-
-public enum Position {
-
-    WORKER,
-    MID_MANAGER,
-    TOP_MANAGER,
-    OWNER
-}
