--- calculator/src/main/java/steammachinist/logging/LoggingAspect.java
@@ -1,46 +1,0 @@
-package steammachinist.logging;
-
-import lombok.extern.slf4j.Slf4j;
-import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.annotation.*;
-import org.springframework.stereotype.Component;
-
-@Aspect
-@Component
-@Slf4j
-public class LoggingAspect {
-
-    @Pointcut("execution(* steammachinist.controller.*.*(..))")
-    public void controllerMethods() {}
-
-    @Pointcut("execution(* steammachinist.service.CalculatorService.prescoreAndGenerateOffers(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.generateIncompletePossibleOffers(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.calculateOffer*(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.scoreAndCalculateCredit(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.generateInitialCredit(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.scoreCreditRate(..)) " +
-            "|| execution(* steammachinist.service.CalculatorService.calculateCredit*(..))")
-    public void serviceMethods() {}
-
-    @Pointcut("controllerMethods() || serviceMethods()")
-    public void combinedPointcut() {}
-
-    @Before("combinedPointcut()")
-    public void logMethodCall(JoinPoint joinPoint) {
-        String methodName = joinPoint.getSignature().getName();
-        Object[] args = joinPoint.getArgs();
-        log.info("Method {} called with arguments: {}", methodName, args);
-    }
-
-    @AfterReturning(pointcut = "combinedPointcut()", returning = "result")
-    public void logMethodReturn(JoinPoint joinPoint, Object result) {
-        String methodName = joinPoint.getSignature().getName();
-        log.info("Method {} returned with result: {}", methodName, result);
-    }
-
-    @AfterThrowing(pointcut = "combinedPointcut()", throwing = "exception")
-    public void logMethodException(JoinPoint joinPoint, Throwable exception) {
-        String methodName = joinPoint.getSignature().getName();
-        log.error("Method {} thrown exception: {}", methodName, exception.getMessage());
-    }
-}
