--- calculator/src/main/java/steammachinist/validation/AdultValidator.java
@@ -1,21 +1,0 @@
-package steammachinist.validation;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-import java.time.LocalDate;
-import java.time.Period;
-
-public class AdultValidator implements ConstraintValidator<Adult, LocalDate> {
-
-    @Override
-    public void initialize(Adult constraintAnnotation) {
-    }
-
-    @Override
-    public boolean isValid(LocalDate birthDate, ConstraintValidatorContext context) {
-        if (birthDate == null) {
-            return false;
-        }
-        return Period.between(birthDate, LocalDate.now()).getYears() >= 18;
-    }
-}
