--- calculator/src/test/java/steammachinist/CalculatorApplicationTest.java
@@ -1,12 +1,0 @@
-package steammachinist;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class CalculatorApplicationTest {
-
-    @Test
-    void contextLoads() {
-    }
-}
